requirements.txt
asgiref==3.8.1
celery==5.3.6
Django>=3.2
Pillow
django-redis
djangorestframework
django-filter
drf-spectacular
channels
python-decouple
dj-database-url
djangorestframework-simplejwt
drf-yasg==1.21.7
python-dotenv==1.0.1
redis==5.0.4
requests==2.32.3
sqlparse==0.5.0
gunicorn==22.0.0
uvicorn==0.30.1
daphne==4.1.2
